refactor: Move theme and palette into @chainlit/components
