+++ libs/components/theme/index.ts
@@ -1,0 +1,2 @@
+export * from './palette';
+export * from './theme';
