+++ libs/components/theme/palette.ts
